move toaster components to the toaster directory
