+++ morpheus-admin/components/organisms/Toaster/use-toast.ts
@@ -1,7 +1,7 @@
 // Inspired by react-hot-toast library
 import * as React from "react";
 
-import type { ToastActionElement, ToastProps } from "@/components/organisms/toast";
+import type { ToastActionElement, ToastProps } from "@/components/organisms/Toaster/toast";
 
 const TOAST_LIMIT = 5;
 const TOAST_REMOVE_DELAY = 1000000;
